% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_binseg.R,
%   R/changepoint_binseg_mean.R
\name{changepoint_binseg_mean}
\alias{changepoint_binseg_mean}
\title{Changepoint identification with Binary Segmentation algorithm}
\usage{
changepoint_binseg_mean(variable, time, change_points, variance)

changepoint_binseg_mean(variable, time, change_points, variance)
}
\arguments{
\item{variable}{(num) The outcome variable.}

\item{time}{(num) The time point(s) in which the outcome \code{variable}
is measured.}

\item{change_points}{(num) Single number of changepoint that
are expecting}

\item{variance}{(char) if TRUE, then identifies the changepoints
of variance of the outcome \code{variable} (default is considering the
mean).}
}
\value{
(num) vector of length \code{change_points} reporting the
timepoints when the changes happens.

(num) vector of length \code{change_points} reporting the
timepoints when the changes happens.
}
\description{
Identify the changepoints, where the mean or variance of the
response variable changes  using Binary Segmentation algorithm.

Identify the changepoints, where the mean or variance of the
response variable changes  using Binary Segmentation algorithm.
}
\examples{
# Prepare a dummy data set
variable <- rnorm(100, 50, 5)
time <- c(seq(1,30), seq(1, 40), seq(1,20), seq(1,10))
change_points <- 3

# Identify the 3 change points of the mean of variable using BinSeg
# algorithm
changepoint_binseg(variable, time, change_points)
# Identify the 3 change points of the variance of variable using
# BinSeg algorithm
changepoint_binseg(variable, time, change_points, variance = TRUE)
# Prepare a dummy data set
variable <- rnorm(100, 50, 5)
time <- c(seq(1,30), seq(1, 40), seq(1,20), seq(1,10))
change_points <- 3

# Identify the 3 change points of the mean of variable using BinSeg
# algorithm
changepoint_binseg_mean(variable, time, change_points)
}
