% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_binseg_var.R
\name{changepoint_binseg_var}
\alias{changepoint_binseg_var}
\title{Changepoint BinSeg algorithm using the variance function}
\usage{
changepoint_binseg_var(variable, time, change_points)
}
\arguments{
\item{variable}{(num) The outcome variable.}

\item{time}{(num) The time point(s) in which the outcome \code{variable}
is measured.}

\item{change_points}{(num) Single number of changepoint that
are expecting}
}
\value{
(num) vector of length \code{change_points} reporting the
timepoints when the changes happens.
}
\description{
Identify the changepoints where the variability of the outcome
variable changes using Binary Segmentation algorithm.
}
\examples{
# Prepare a dummy data set
variable <- rnorm(100, 50, 5)
time <- c(seq(1,30), seq(1, 40), seq(1,20), seq(1,10))
change_points <- 3

# Identify the 3 change points of the mean of variable using BinSeg
# algorithm
changepoint_binseg_var(variable, time, change_points)
}
