% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_pelt.R
\name{changepoint_pelt}
\alias{changepoint_pelt}
\title{Change point identification with PELT algorithm}
\usage{
changepoint_pelt(variable, time, stat = c("mean", "variance"))
}
\arguments{
\item{variable}{(num) The outcome variable.}

\item{time}{(num) The time point(s) in which the outcome
\code{variable} is measured.}

\item{stat}{(char) if variance, then identifies the time points, when
the variance of the outcome \code{variable} changes.
(default is considering the mean).}
}
\value{
(num) vector of length \code{change_points} reporting the
time points when the changes happen.
}
\description{
Identify the time points, when the mean or variance of the
response variable changes using the Pruned Exact Linear Time
\code{(PELT)} algorithm.
}
\examples{
# Prepare a dummy data set
variable <- rnorm(100, 50, 5)
time <- c(seq(1,30), seq(1, 40), seq(1,20), seq(1,10))

# Identify the change points of the mean of variable using PELT
# algorithm
 changepoint_pelt(variable, time)
# Identify the change points of the variance of variable using
# PELT algorithm
 changepoint_pelt(variable, time, stat = "variance")
}
