% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_pelt_mean.R
\name{changepoint_pelt_mean}
\alias{changepoint_pelt_mean}
\title{Change point identification with PELT algorithm}
\usage{
changepoint_pelt_mean(variable, time)
}
\arguments{
\item{variable}{(num) The outcome variable.}

\item{time}{(num) The time point(s) in which the outcome \code{variable}
is measured.}
}
\value{
(num) vector of length \code{change_points} reporting the
time points when the changes happen.
}
\description{
Identify the change points, where the mean of the response variable
changes  using the Pruned Exact Linear Time \code{(PELT)}  algorithm.
}
\examples{
# Prepare a dummy data set
variable <- rnorm(100, 50, 5)
time <- c(seq(1,30), seq(1, 40), seq(1,20), seq(1,10))

# Identify the change points of the mean of variable using pelt
# algorithm
changepoint_pelt_mean(variable, time)
}
