% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_pelt_var.R
\name{changepoint_pelt_var}
\alias{changepoint_pelt_var}
\title{Changepoint identification with PELT algorithm}
\usage{
changepoint_pelt_var(variable, time)
}
\arguments{
\item{variable}{(num) The outcome variable.}

\item{time}{(num) The time point(s) in which the outcome \code{variable}
is measured.}
}
\value{
(num) vector of length \code{change_points} reporting the
timepoints when the changes happens.
}
\description{
Identify the changepoints, where the variance of the response
variable changes  using the Pruned Exact Linear Time \code{(PELT)}
algorithm.
}
\examples{
# Prepare a dummy data set
variable <- rnorm(100, 50, 5)
time <- c(seq(1,30), seq(1, 40), seq(1,20), seq(1,10))

# Identify the change points of the variance of variable using pelt
# algorithm
changepoint_pelt_var(variable, time)
}
