% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varifinder-package.R
\docType{package}
\name{varifinder-package}
\alias{varifinder}
\alias{varifinder-package}
\title{varifinder: Variability Ratio Index Finder}
\description{
Calculate VARI to compare the change points in retrospective data.
}
\author{
\strong{Maintainer}: Noor muhammad Khan \email{noor.muhammadkhan@studenti.unipd.it} (\href{https://orcid.org/0000-0003-4713-7277}{ORCID})

}
\keyword{internal}
