% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varifinder.R
\name{varifinder}
\alias{varifinder}
\title{Variability Ratio Index (VARI)}
\usage{
varifinder(variable, phase, time)
}
\arguments{
\item{variable}{(num) The outcome variable in long format.}

\item{phase}{(character) list of clinical phases of the total time
in long format.}

\item{time}{(num) vector of time points, when the outcome is measured
in long format.}
}
\value{
(num) data frame containing rows for the unique phases.
The column \code{Times} presents the total number of time points within
that phase. \code{CPT_M} and \code{CPT_V} present the total number of
detected change points in each phase based on mean and variance
function of the variable respectively. Finally, the column \code{VARI_M}
and \code{VARI_V} present the variability ratio index for each phase.
}
\description{
Calculate the Variability Ratio Index (VARI) to compare the number of
detected change points between different clinical phases in
retrospective data.
}
\examples{
# Prepare a dummy data set
variable <- rnorm(100, 50, 5)
time <- c(seq(1,30), seq(1, 40), seq(1,20), seq(1,10))
phase <- rep(factor(c(1, 2, 3, 4)), c(30, 40, 20, 10))

# Identify the VARI
 varifinder(variable, phase, time)
}
