% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varifinder_ci.R
\name{varifinder_ci}
\alias{varifinder_ci}
\title{Confidence Interval of the Variability Ratio Index.}
\usage{
varifinder_ci(time, cpt, vari)
}
\arguments{
\item{time}{(num) The time point(s) in which the outcome variable
is measured.}

\item{cpt}{(num) The number of change points detected within each
phase.}

\item{vari}{(num) Variability Ratio Index.}
}
\value{
(num) data frame containing rows for the unique phases.
The column \code{Times} presents the total number of time points in
within  that phase. \code{VARI} presents the Variability
Ratio Index for each phase. The \code{Lower_limit} and the \code{Upper_limit}
presents the 95\% exact confidence interval for the VARI.
}
\description{
Identify the 95\% exact confidence interval for the Variability Ratio
Index (VARI).
}
\examples{
# Prepare a dummy data set
variable <- rnorm(100, 50, 5)
time <- c(seq(1,30), seq(1, 40), seq(1,20), seq(1,10))
phase <- rep(factor(c(1, 2, 3, 4)), c(30, 40, 20, 10))


# Identify the 95\% exact confidence interval for VARI when it is
# calculated using mean function
vari <- varifinder(variable, phase, time)
varifinder_ci(vari$Time, vari$CPT_M, vari$VARI_M)

# Identify the 95\% exact confidence interval for VARI when it is
# calculated using variance function
varifinder_ci(vari$Time, vari$CPT_V, vari$VARI_V)
}
