% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/changepoint_CI.R
\name{changepoint_CI}
\alias{changepoint_CI}
\title{Confidence Interval of mean difference at the Changepoint}
\usage{
changepoint_CI(variable, changepoints)
}
\arguments{
\item{variable}{(num) The outcome variable.}

\item{changepoints}{(num) vector of length \code{change_points} reporting
the timepoints when the changes happens.}
}
\value{
(num) data frame including amount of mean differences at
every change points with 95\% confidence interval.
}
\description{
Calculate the 95\% confidence interval for the mean differences
at the changepoints.
}
\examples{
# Prepare a dummy data set
variable <- rnorm(100, 50, 5)
time <- c(seq(1,30), seq(1, 40), seq(1,20), seq(1,10))
change_points <- 3

# Identify the change points of the mean of variable using BinSeg
# algorithm
 cpt_binseg <- changepoint_binseg(variable, time, change_points)
 changepoint_CI(variable, cpt_binseg)
# Identify the change points of the mean of variable using
# PELT algorithm
 cpt_pelt <- changepoint_pelt(variable, time)
 changepoint_CI(variable, cpt_pelt)
}
