% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varifinder.R
\name{varifinder}
\alias{varifinder}
\title{Variability Ratio Index (VARI)}
\usage{
varifinder(variable, phase, time)
}
\arguments{
\item{variable}{(num) The outcome variable in long format.}

\item{phase}{(character) list of clinical phases of the total time
in long format.}

\item{time}{(num) vector of time points, when the outcome is measured
in long format.}
}
\value{
(num) data frame containing rows for the the unique phases.
Times points, change points and VARI based on mean and variance for
each phase.
}
\description{
Calculate the Variability Ratio Index (VARI) index.
}
\examples{
# Prepare a dummy data set
variable <- rnorm(100, 50, 5)
time <- c(seq(1,30), seq(1, 40), seq(1,20), seq(1,10))
phase <- rep(factor(c(1, 2, 3, 4)), c(30, 40, 20, 10))

# Identify the VARI
 varifinder(variable, phase, time)
}
